refactor: spread the contents of helpers to multiple layers
