+++ src/features/hints/ui/HintControls.ts
@@ -2,7 +2,7 @@ import { Component, i, input, label } from "../../../shared/ui";
 import { HintSettings, HintSettingsData } from "../model/HintSettings";
 
 import { Observer, Publisher } from "../../../entities/state";
-import { isValidSetting } from "../../../shared/helpers";
+import { isValidSetting } from "../../../shared/lib";
 
 import styles from "./HintControls.module.css";
