+++ src/features/levels/config/constants.ts
@@ -1,0 +1,5 @@
+export const RATING_THRESHOLDS = {
+  perfect: 1,
+  good: 0.7,
+  passed: 0.5,
+};
