+++ src/features/levels/ui/LevelControls.ts
@@ -1,5 +1,5 @@
 import { Component, label, option, select } from "../../../shared/ui";
-import { isValidSetting } from "../../../shared/helpers";
+import { isValidSetting } from "../../../shared/lib";
 import { Observer, Publisher } from "../../../entities/state";
 
 import { LevelsState } from "../model/LevelsState";
