+++ src/shared/api/urls.ts
@@ -1,0 +1,4 @@
+export const BASE_URL =
+  "https://github.com/rolling-scopes-school/rss-puzzle-data/raw/main";
+export const IMAGES_BASE_URL =
+  "https://raw.githubusercontent.com/rolling-scopes-school/rss-puzzle-data/main/images";
