+++ src/shared/lib/assertNonNull.ts
@@ -1,0 +1,6 @@
+export function assertNonNull<T>(value: T | null | undefined): T {
+  if (value === null || value === undefined) {
+    throw new Error(`Not defined!`);
+  }
+  return value;
+}
