+++ src/shared/lib/isValidSetting.ts
@@ -1,0 +1,6 @@
+export function isValidSetting<T extends object>(
+  key: string | number | symbol,
+  state: T,
+): key is keyof T {
+  return key in state;
+}
