+++ src/widgets/game-fields/config/constants.ts
@@ -1,0 +1,1 @@
+export const ANIMATION_DELAY_COEFFICIENT = 50;
