+++ src/widgets/game-fields/ui/words-picker/WordsPicker.ts
@@ -3,10 +3,8 @@ import { PickRow } from "../rows/PickRow";
 import { RoundState } from "../../../../features/levels";
 import { HintSettings } from "../../../../features/hints";
 import { Observer, Publisher } from "../../../../entities/state";
-import {
-  calculateImageAspectRatio,
-  debounceListener,
-} from "../../../../shared/helpers";
+import { calculateImageAspectRatio } from "../../lib/calculateImageAspectRatio";
+import { debounceListener } from "../../../../shared/lib";
 
 import styles from "./WordsPicker.module.css";
